refactor: checkInGallery react query 전환 및 모달 분리
